move util/variable to single point of use
