+++ pkg/cmd/openshift-controller-manager/imageformat/imagetemplate.go
@@ -1,4 +1,4 @@
-package variable
+package imageformat
 
 import (
 	"fmt"
