+++ pkg/cmd/openshift-controller-manager/imageformat/variable_test.go
@@ -1,4 +1,4 @@
-package variable
+package imageformat
 
 import (
 	"testing"
